FACES-1543 Create ApplicantViewBean for commentRendered property in job applicant demo portlets
