+++ demos/bridge/icefaces3-compat-portlet/src/main/java/com/liferay/faces/demos/bean/ApplicantBackingBean.java
@@ -18,7 +18,7 @@ import java.util.List;
 
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ManagedProperty;
-import javax.faces.bean.ViewScoped;
+import javax.faces.bean.RequestScoped;
 import javax.faces.component.UIComponent;
 import javax.faces.component.UIInput;
 import javax.faces.context.FacesContext;
@@ -43,7 +43,7 @@ import com.liferay.faces.util.logging.LoggerFactory;
  * @author  "Neil Griffin"
  */
 @ManagedBean(name = "applicantBackingBean")
-@ViewScoped
+@RequestScoped
 public class ApplicantBackingBean implements Serializable {
 
 	// serialVersionUID
@@ -62,7 +62,6 @@ public class ApplicantBackingBean implements Serializable {
 	private transient ListModelBean listModelBean;
 
 	// JavaBeans Properties for UI
-	private boolean commentsRendered = false;
 	private String fileUploadAbsolutePath;
 	private boolean popupRendered;
 	private String uploadedFileId;
@@ -146,7 +145,7 @@ public class ApplicantBackingBean implements Serializable {
 
 			if (city != null) {
 				applicantModelBean.setAutoFillCity(city.getCityName());
-				
+
 				// begin fix for FACES-1537
 				UIComponent postalCodeInputComponent = valueChangeEvent.getComponent();
 				UIInput cityInputComponent = (UIInput) postalCodeInputComponent.findComponent("city");
@@ -154,7 +153,7 @@ public class ApplicantBackingBean implements Serializable {
 				cityInputComponent.setValid(true);
 				cityInputComponent.setValue(null);
 				// end fix for FACES-1537
-				
+
 				applicantModelBean.setAutoFillProvinceId(city.getProvinceId());
 			}
 		}
@@ -217,10 +216,6 @@ public class ApplicantBackingBean implements Serializable {
 		}
 	}
 
-	public void toggleComments(ActionEvent actionEvent) {
-		commentsRendered = !commentsRendered;
-	}
-
 	public void togglePopup(ActionEvent actionEvent) {
 		popupRendered = !popupRendered;
 	}
@@ -229,14 +224,6 @@ public class ApplicantBackingBean implements Serializable {
 
 		// Injected via @ManagedProperty annotation
 		this.applicantModelBean = applicantModelBean;
-	}
-
-	public void setCommentsRendered(boolean commentsRendered) {
-		this.commentsRendered = commentsRendered;
-	}
-
-	public boolean isCommentsRendered() {
-		return commentsRendered;
 	}
 
 	public boolean isPopupRendered() {
