+++ demos/bridge/icefaces3-portlet/src/main/java/com/liferay/faces/demos/bean/ApplicantViewBean.java
@@ -1,0 +1,48 @@
+/**
+ * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+package com.liferay.faces.demos.bean;
+
+import java.io.Serializable;
+
+import javax.faces.bean.ManagedBean;
+import javax.faces.bean.ViewScoped;
+
+
+/**
+ * This is a JSF view managed-bean for the applicant.xhtml composition.
+ *
+ * @author  Kyle Stiemann
+ */
+@ManagedBean(name = "applicantViewBean")
+@ViewScoped
+public class ApplicantViewBean implements Serializable {
+
+	// serialVersionUID
+	private static final long serialVersionUID = 5456293483676149659L;
+
+	// JavaBean Properties for UI
+	private boolean commentsRendered = false;
+
+	public void toggleComments() {
+		commentsRendered = !commentsRendered;
+	}
+
+	public void setCommentsRendered(boolean commentsRendered) {
+		this.commentsRendered = commentsRendered;
+	}
+
+	public boolean isCommentsRendered() {
+		return commentsRendered;
+	}
+}
