+++ demos/bridge/jsf2-jsp-portlet/src/main/java/com/liferay/faces/demos/bean/ApplicantBackingBean.java
@@ -19,7 +19,7 @@ import java.util.List;
 
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ManagedProperty;
-import javax.faces.bean.ViewScoped;
+import javax.faces.bean.RequestScoped;
 import javax.faces.context.FacesContext;
 import javax.faces.event.ActionEvent;
 import javax.faces.event.ValueChangeEvent;
@@ -38,7 +38,7 @@ import com.liferay.faces.util.logging.LoggerFactory;
  * @author  "Neil Griffin"
  */
 @ManagedBean(name = "applicantBackingBean")
-@ViewScoped
+@RequestScoped
 public class ApplicantBackingBean implements Serializable {
 
 	// serialVersionUID
@@ -54,7 +54,6 @@ public class ApplicantBackingBean implements Serializable {
 	private transient ListModelBean listModelBean;
 
 	// JavaBeans Properties for UI
-	private boolean commentsRendered = false;
 	private boolean fileUploaderRendered = false;
 
 	private transient HtmlInputFile attachment1;
@@ -161,10 +160,6 @@ public class ApplicantBackingBean implements Serializable {
 		}
 	}
 
-	public void toggleComments(ActionEvent actionEvent) {
-		commentsRendered = !commentsRendered;
-	}
-
 	@SuppressWarnings("deprecation")
 	public void uploadAttachments(ActionEvent actionEvent) {
 
@@ -221,14 +216,6 @@ public class ApplicantBackingBean implements Serializable {
 
 	public void setAttachment3(HtmlInputFile attachment3) {
 		this.attachment3 = attachment3;
-	}
-
-	public void setCommentsRendered(boolean commentsRendered) {
-		this.commentsRendered = commentsRendered;
-	}
-
-	public boolean isCommentsRendered() {
-		return commentsRendered;
 	}
 
 	public boolean isFileUploaderRendered() {
