+++ demos/bridge/jsf2-jsp-portlet/src/main/webapp/views/applicant.jspx
@@ -49,8 +49,8 @@
 							<h:message id="provinceIdMessage" for="provinceId" />
 						</aui:field>
 						<aui:field id="postalCodeField" label="#{i18n['zip-postal']}">
-							<h:inputText id="postalCode" immediate="true" onchange="this.form.submit()" required="true" value="#{applicantModelBean.postalCode}"
-								valueChangeListener="#{applicantBackingBean.postalCodeListener}" />
+							<h:inputText id="postalCode" immediate="true" onchange="this.form.submit()" required="true"
+								value="#{applicantModelBean.postalCode}" valueChangeListener="#{applicantBackingBean.postalCodeListener}" />
 							<h:graphicImage library="example" name="icon-help.png" title="#{i18n['postal-code-hint']}" />
 							<h:message id="postalCodeMessage" for="postalCode" />
 						</aui:field>
@@ -60,12 +60,12 @@
 					<aui:column>
 						<aui:field id="commentsField" label="#{i18n['comments']}">
 							<h:panelGroup id="showHide" layout="block">
-								<h:commandLink actionListener="#{applicantBackingBean.toggleComments}" immediate="true"
-									rendered="#{!applicantBackingBean.commentsRendered}" value="#{i18n['show']} #{i18n['comments']}" />
-								<h:commandLink actionListener="#{applicantBackingBean.toggleComments}" immediate="true"
-									rendered="#{applicantBackingBean.commentsRendered}" value="#{i18n['hide']} #{i18n['comments']}" />
+								<h:commandLink actionListener="#{applicantViewBean.toggleComments}" immediate="true"
+									rendered="#{!applicantViewBean.commentsRendered}" value="#{i18n['show']} #{i18n['comments']}" />
+								<h:commandLink actionListener="#{applicantViewBean.toggleComments}" immediate="true"
+									rendered="#{applicantViewBean.commentsRendered}" value="#{i18n['hide']} #{i18n['comments']}" />
 							</h:panelGroup>
-							<h:inputTextarea id="comments" cols="50" rows="10" rendered="#{applicantBackingBean.commentsRendered}"
+							<h:inputTextarea id="comments" cols="50" rows="10" rendered="#{applicantViewBean.commentsRendered}"
 								value="#{applicantModelBean.comments}" />
 						</aui:field>
 					</aui:column>
@@ -119,7 +119,9 @@
 			<h:form id="f1" enctype="multipart/form-data">
 				<h:messages />
 				<h:panelGrid columns="1">
-					<h3><h:outputText value="#{i18n['attachments']}" /></h3>
+					<h3>
+						<h:outputText value="#{i18n['attachments']}" />
+					</h3>
 					<bridge:inputFile binding="#{applicantBackingBean.attachment1}" />
 					<bridge:inputFile binding="#{applicantBackingBean.attachment2}" />
 					<bridge:inputFile binding="#{applicantBackingBean.attachment3}" />
