+++ demos/bridge/richfaces4-portlet/src/main/webapp/views/applicant.xhtml
@@ -82,12 +82,12 @@
 					<aui:column>
 						<aui:field id="commentsField" label="#{i18n['comments']}">
 							<h:panelGroup id="showHide" layout="block">
-								<a4j:commandLink actionListener="#{applicantBackingBean.toggleComments}" immediate="true" render="commentsField"
-									rendered="#{!applicantBackingBean.commentsRendered}" value="#{i18n['show']} #{i18n['comments']}" />
-								<a4j:commandLink actionListener="#{applicantBackingBean.toggleComments}" immediate="true" render="commentsField"
-									rendered="#{applicantBackingBean.commentsRendered}" value="#{i18n['hide']} #{i18n['comments']}" />
+								<a4j:commandLink actionListener="#{applicantViewBean.toggleComments}" immediate="true" render="commentsField"
+									rendered="#{!applicantViewBean.commentsRendered}" value="#{i18n['show']} #{i18n['comments']}" />
+								<a4j:commandLink actionListener="#{applicantViewBean.toggleComments}" immediate="true" render="commentsField"
+									rendered="#{applicantViewBean.commentsRendered}" value="#{i18n['hide']} #{i18n['comments']}" />
 							</h:panelGroup>
-							<h:inputTextarea id="comments" cols="50" rows="10" rendered="#{applicantBackingBean.commentsRendered}"
+							<h:inputTextarea id="comments" cols="50" rows="10" rendered="#{applicantViewBean.commentsRendered}"
 								value="#{applicantModelBean.comments}">
 								<f:ajax />
 							</h:inputTextarea>
